@@ -1,33 +1,32 @@
 {
-  "name": "invert-kv",
-  "version": "1.0.0",
-  "description": "Invert the key/value of an object. Example: {foo: 'bar'} → {bar: 'foo'}",
-  "license": "MIT",
-  "repository": "sindresorhus/invert-kv",
-  "author": {
-    "name": "Sindre Sorhus",
-    "email": "sindresorhus@gmail.com",
-    "url": "http://sindresorhus.com"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  },
-  "scripts": {
-    "test": "mocha"
-  },
-  "files": [
-    "index.js"
-  ],
-  "keywords": [
-    "object",
-    "obj",
-    "key",
-    "value",
-    "val",
-    "kv",
-    "invert"
-  ],
-  "devDependencies": {
-    "mocha": "*"
-  }
+	"name": "invert-kv",
+	"version": "1.0.0",
+	"description": "Invert the key/value of an object. Example: `{foo: 'bar'}` → `{bar: 'foo'}`",
+	"license": "MIT",
+	"repository": "sindresorhus/invert-kv",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "sindresorhus.com"
+	},
+	"engines": {
+		"node": ">=4"
+	},
+	"scripts": {
+		"test": "xo && ava"
+	},
+	"files": [
+		"index.js"
+	],
+	"keywords": [
+		"object",
+		"key",
+		"value",
+		"kv",
+		"invert"
+	],
+	"devDependencies": {
+		"ava": "*",
+		"xo": "*"
+	}
 }
